backend/requirements.txt
FastAPI==0.115.0
uvicorn>=0.27.0
python-multipart==0.0.6
python-dotenv==1.0.0
Pillow==10.1.0
pytesseract==0.3.10
opencv-python==4.8.1.78
openpyxl==3.1.2
transformers==4.35.2
torch>=2.2.0
scikit-learn==1.3.2
pandas==2.1.3
numpy==1.26.2
requests==2.31.0
celery==5.3.4
redis==5.0.1
pydantic==2.8.0
pydantic-settings==2.2.0
pymongo==4.6.1
motor==3.3.2
Werkzeug==3.0.1
supabase==2.3.4
PyJWT==2.10.1
python-jose==3.3.0
httpx==0.25.2
paddlepaddle-gpu==2.6.2
paddleocr==2.8.1
